Closes #19

+++ src/AppBundle/Controller/StockController.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\Stocks;
+use AppBundle\Entity\StocksStockMarket;
 use AppBundle\Form\StocksType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -20,8 +21,9 @@ class StockController extends BasicController
 
         $this->getModes();
 
-        $repo_typeStocks = $this->getDoctrine()->getRepository('AppBundle:TypeStocks');
         $repo_stocks = $this->getDoctrine()->getRepository('AppBundle:Stocks');
+
+        $repo_typeStocks = $this->getDoctrine()->getRepository('AppBundle:TypeStocks');
 
         $typeDraft = $repo_typeStocks->returnType('Fût');
         $typeBottle = $repo_typeStocks->returnType('Bouteille');
@@ -57,6 +59,11 @@ class StockController extends BasicController
         $this->getModes();
 
         // 1) Récupérer l'Article et construire le form
+        if (in_array("stockmarket", $this->data['activeModes'])) {
+            $this->addFlash('error', "Cette action n'est pas autorisée en mode Stock Market !");
+            return $this->redirectToRoute('stock');
+        }
+
         $repo_stocks = $this->getDoctrine()->getRepository('AppBundle:Stocks');
 
         $article = $repo_stocks->find($id_article);
@@ -83,7 +90,7 @@ class StockController extends BasicController
         }
 
         $this->data['form'] = $form->createView();
-        $this->data['mode'] = 'modify_article';
+        $this->data['form_mode'] = 'modify_article';
         $this->data['nom'] = $article->getNom();
         $this->data['type'] = $type;
 
@@ -106,7 +113,13 @@ class StockController extends BasicController
         $this->getModes();
 
         // 1) Construire le form
+        if (in_array("stockmarket", $this->data['activeModes'])) {
+            $this->addFlash('error', "Cette action n'est pas autorisée en mode Stock Market !");
+            return $this->redirectToRoute('stock');
+        }
+
         $article = new Stocks();
+
         $form = $this->createForm(StocksType::class, $article);
 
         // 2) Traiter le submit (uniquement sur POST)
@@ -126,7 +139,7 @@ class StockController extends BasicController
         }
 
         $this->data['form'] = $form->createView();
-        $this->data['mode'] = 'new_article';
+        $this->data['form_mode'] = 'new_article';
 
         return $this->render(
             'stock/article.html.twig',
@@ -144,18 +157,31 @@ class StockController extends BasicController
             throw $this->createAccessDeniedException();
         }
 
-        $idarticle = $request->query->get('idarticle');
+        $this->getModes();
 
-        $em = $this->getDoctrine()->getManager();
-        $article = $em->getRepository('AppBundle:Stocks')->find($idarticle);
-
-        if (!$article) {
-            throw $this->createNotFoundException("Article non trouvé pour l'id " . $idarticle);
+        if (in_array("stockmarket", $this->data['activeModes'])) {
+            $this->addFlash('error', "Cette action n'est pas autorisée en mode Stock Market !");
+            return $this->redirectToRoute('stock');
         }
 
-        $transactions = $em->getRepository('AppBundle:DetailsTransactions')->findOneBy(['article' => $idarticle]);
-        foreach ($transactions as $elt_commande) {
-            $em->remove($elt_commande);
+        $idArticle = $request->query->get('idarticle');
+        $em = $this->getDoctrine()->getManager();
+        $article = $em->getRepository('AppBundle:Stocks')->find($idArticle);
+
+        if (!$article) {
+            throw $this->createNotFoundException("Article non trouvé pour l'id " . $idArticle);
+        }
+
+        $transactions = $em->getRepository('AppBundle:DetailsTransactions')->findBy(['article' => $idArticle]);
+        if (isset($transactions)) {
+            foreach ($transactions as $elt_commande) {
+                $em->remove($elt_commande);
+            }
+        }
+
+        $sm_data = $em->getRepository('AppBundle:StockMarketData')->find($idArticle);
+        if (isset($sm_data)) {
+            $em->remove($sm_data);
         }
 
         $em->remove($article);
